Limiting samples approach, lightning fast

+++ bit_autocorrelation.cpp
@@ -2,8 +2,8 @@
 
 constexpr int BUFFER_SIZE = 2048;
 
-std::vector<int16_t> autocorrelate(const std::vector<int16_t>& data){
-    std::vector<int16_t> result = {};
+std::vector<uint32_t> autocorrelate(const std::vector<int16_t>& data){
+    std::vector<uint32_t> result = {};
     const auto size = data.size();
     std::cout << "Data size: " << size << "\n";
     result.push_back(0);
@@ -14,7 +14,6 @@ std::vector<int16_t> autocorrelate(const std::vector<int16_t>& data){
         }
         result.push_back(zero_count);
     }
-
     return result;
 }
 
@@ -25,36 +24,3 @@ std::vector<int16_t> zero_cross(const std::vector<int16_t>& data){
     }
     return result;
 }
-
-// int main() {
-//     std::array<float, BUFFER_SIZE> signal = {};
-//     // Sample_rate/frequency must be less than BUFFER_SIZE/2
-//     int sample_rate = 80000.0f;
-//     float frequency = 250;
-//     float p  = float(sample_rate) / frequency;
-//     for (std::size_t i = 0; i < signal.size(); ++i) {
-//         signal[i] = 0.3 * sin(2 * M_PI * i / p) + 0.4 * sin(4 * M_1_PI * i / p) + 0.3 * sin(6 * M_PI * i / p);
-//     }
-
-//     auto binary_signal = zero_cross(signal);
-//     auto autocorr_result = autocorrelate(binary_signal);
-//     int min_count = BUFFER_SIZE;
-
-//     for (std::size_t i = 50; i < BUFFER_SIZE/2; ++i) {
-//         if (autocorr_result[i] < min_count) {
-//             min_count = autocorr_result[i];
-//         }
-//         std::cout << "Lag " << i << ": " << autocorr_result[i] << "\n";
-//     }
-
-//     constexpr auto mid_pos = BUFFER_SIZE / 2;
-//     for (int i = 50; i < mid_pos; i++) {
-//         if (autocorr_result[i] <= min_count + 2) {
-//             std::cout << "Period at skip: " << i << "\n";
-//             std::cout << "Estimated frequency: " << sample_rate / i << " Hz\n";
-//             break;
-//         }
-//     }
-
-//     return 0;
-// }
